+++ Market.API/Domain/Entities/PriceHistory.cs
@@ -3,10 +3,10 @@
 public class PriceHistory : BaseEntity
 {
     public string Symbol { get; set; } = string.Empty;
-    public double OpenPrice { get; set; }
-    public double ClosePrice { get; set; }
-    public double HighPrice { get; set; }
-    public double LowPrice { get; set; }
-    public double Volume { get; set; }
+    public decimal OpenPrice { get; set; }
+    public decimal ClosePrice { get; set; }
+    public decimal HighPrice { get; set; }
+    public decimal LowPrice { get; set; }
+    public decimal Volume { get; set; }
     public DateTime Timestamp { get; set; }
 }
